feat: доработать под чек-лист
