+++ README.md
@@ -1,2 +1,7 @@
-# movies-explorer-api
-Backend
+# Дипломная работа на курсе веб-разработчик
+Backend приложения с фильмами.
+## Ссылки на проект
+
+IP 89.111.131.44
+
+Backend [https://api.movies.studioit.online](https://api.movies.studioit.online)
