+++ app.js
@@ -1,30 +1,18 @@
+require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 
-const { PORT = 3001, NODE_ENV, DB_ADDRESS } = process.env;
+const { PORT = 3000, NODE_ENV, DB_ADDRESS } = process.env;
 const app = express();
 const { errors } = require('celebrate');
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
-const { NotFoundError } = require('./errors');
-const { loginValidation, createUserValidation } = require('./middlewares/validation');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-
-const {
-  login,
-  createUser,
-} = require('./controllers/users');
-const auth = require('./middlewares/auth');
+const { DB_ADDRESS_LOCAL } = require('./utils/constants');
 const handleError = require('./middlewares/handleError');
-
-const limiter = rateLimit({
-  windowMs: 10 * 60 * 1000,
-  max: 100,
-  standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-});
+const limiter = require('./middlewares/rateLimit');
+const router = require('./routes');
 
 app.use(limiter);
 app.use(helmet());
@@ -33,7 +21,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cookieParser());
 
-const DB_URL = NODE_ENV === 'production' ? DB_ADDRESS : 'mongodb://127.0.0.1:27017/bitfilmsdb';
+const DB_URL = NODE_ENV === 'production' ? DB_ADDRESS : DB_ADDRESS_LOCAL;
 mongoose.connect(DB_URL, {
   useNewUrlParser: true,
 })
@@ -42,16 +30,7 @@ mongoose.connect(DB_URL, {
 
 app.use(requestLogger); // подключаем логгер запросов
 
-app.post('/signin', loginValidation, login);
-app.post('/signup', createUserValidation, createUser);
-
-app.use(auth);
-app.use('/users', require('./routes/users'));
-app.use('/movies', require('./routes/movies'));
-
-app.all('*', () => {
-  throw new NotFoundError('Такого не существует(');
-});
+app.use(router);
 
 app.use(errorLogger); // подключаем логгер ошибок
